requirements.txt
# Core Dependencies
python-dotenv>=1.0.0
pydantic>=2.5.0

# Document Processing
pypdf2>=3.0.1
python-docx>=1.1.0
langchain>=0.1.0
langchain-community>=0.0.13

# Embeddings
sentence-transformers>=2.2.2
torch>=2.0.0

# Vector Database
pinecone-client>=3.0.0

# LLM - Ollama for local LLM (Qwen 2.5)
ollama>=0.1.6

# UI
streamlit>=1.29.0

# Utilities
tqdm>=4.66.1
numpy>=1.24.3
